backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.127.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.40.0",
]
